added new error type + renamed reused form components
